blog_to_podcast/requirements.txt
fastapi>=0.115.0
groq>=0.11.0
imageio>=2.37.0
imageio-ffmpeg>=0.6.0
langchain>=0.3.8
langchain-community>=0.3.8
langchain-core>=0.3.18
murf>=2.2.4
numpy>=2.1.0
pydub>=0.25.1
pyloudnorm>=0.1.0
python-dotenv>=1.0.1
requests>=2.32.3
tenacity>=8.2.0
trafilatura>=1.9.0
uvicorn[standard]>=0.30.0

